+++ lib/ReactViews/FeatureInfo/FeatureInfoSection.jsx
@@ -95,10 +95,10 @@ export const FeatureInfoSection = observer(
           .filter(col => col.name && col.title && col.name !== col.title)
           .map(col => [col.name, col.title]);
       }
-      // Only assign alias if it is undefined
-      aliases.forEach(aliasMap => {
-        propertyData[aliasMap[0]] =
-          propertyData[aliasMap[0]] ?? propertyData[aliasMap[1]];
+
+      // Always overwrite using aliases so that applying titles to columns doesn't break feature info templates
+      aliases.forEach(([name, title]) => {
+        propertyData[name] = propertyData[title];
       });
 
       // Properties accessible as {name, value} array; useful when you want
@@ -154,16 +154,7 @@ export const FeatureInfoSection = observer(
       const { t } = this.props;
       const template = this.props.template;
       const templateData = this.getTemplateData();
-      // If property names were changed, let the template access the original property names too.
-      if (
-        defined(templateData) &&
-        defined(templateData._terria_columnAliases)
-      ) {
-        for (let i = 0; i < templateData._terria_columnAliases.length; i++) {
-          const alias = templateData._terria_columnAliases[i];
-          templateData[alias.id] = templateData[alias.name];
-        }
-      }
+
       // templateData may not be defined if a re-render gets triggered in the middle of a feature updating.
       // (Recall we re-render whenever feature.definitionChanged triggers.)
       if (defined(templateData)) {
@@ -802,8 +793,8 @@ function getTimeSeriesChartContext(catalogItem, feature, getChartDetails) {
     if (chartDetails) {
       const result = {
         ...chartDetails,
-        id: featureId.replace(/\"/g, ""),
-        data: csvData.replace(/\\n/g, "\\n")
+        id: featureId?.replace(/\"/g, ""),
+        data: csvData?.replace(/\\n/g, "\\n")
       };
       const idAttr = 'id="' + result.id + '" ';
       const sourceAttr = 'sources="1"';