+++ lib/ReactViews/Map/Panels/SharePanel/SharePanel.jsx
@@ -23,6 +23,10 @@ import {
 import PrintView from "./PrintView";
 import Styles from "./share-panel.scss";
 import StorySharePanel from "./StorySharePanel";
+import {
+  Category,
+  ShareAction
+} from "../../../../Core/AnalyticEvents/analyticEvents";
 
 const SharePanel = observer(
   createReactClass({
@@ -361,7 +365,7 @@ const SharePanel = observer(
     },
 
     renderContentForStoryShare() {
-      const { t } = this.props;
+      const { t, terria } = this.props;
       return (
         <Choose>
           <When condition={this.state.shareUrl === ""}>
@@ -375,6 +379,13 @@ const SharePanel = observer(
                 source={this.getShareUrlInputStory("light")}
                 id="share-url"
                 rounded
+                onCopy={text =>
+                  terria.analytics?.logEvent(
+                    Category.share,
+                    ShareAction.storyCopy,
+                    text
+                  )
+                }
               />
               {this.renderWarning()}
             </div>
@@ -384,7 +395,7 @@ const SharePanel = observer(
     },
 
     renderContentForCatalogShare() {
-      const { t } = this.props;
+      const { t, terria } = this.props;
       return (
         <Choose>
           <When condition={this.state.shareUrl === ""}>
@@ -397,6 +408,13 @@ const SharePanel = observer(
                 text={this.state.shareUrl}
                 source={this.getShareUrlInput("light")}
                 id="share-url"
+                onCopy={text =>
+                  terria.analytics?.logEvent(
+                    Category.share,
+                    ShareAction.catalogCopy,
+                    text
+                  )
+                }
               />
               {this.renderWarning()}
             </div>
@@ -406,7 +424,7 @@ const SharePanel = observer(
     },
 
     renderContentWithPrintAndEmbed() {
-      const { t } = this.props;
+      const { t, terria } = this.props;
       const iframeCode = this.state.shareUrl.length
         ? `<iframe style="width: 720px; height: 600px; border: none;" src="${this.state.shareUrl}" allowFullScreen mozAllowFullScreen webkitAllowFullScreen></iframe>`
         : "";
@@ -414,7 +432,17 @@ const SharePanel = observer(
       return (
         <div>
           <div className={DropdownStyles.section}>
-            <Clipboard source={this.getShareUrlInput("dark")} id="share-url" />
+            <Clipboard
+              source={this.getShareUrlInput("dark")}
+              id="share-url"
+              onCopy={text =>
+                terria.analytics?.logEvent(
+                  Category.share,
+                  ShareAction.shareCopy,
+                  text
+                )
+              }
+            />
             {this.renderWarning()}
           </div>
           <div className={DropdownStyles.section}>