+++ lib/Table/createLongitudeLatitudeFeaturePerRow.ts
@@ -1,16 +1,16 @@
 import Cartesian3 from "terriajs-cesium/Source/Core/Cartesian3";
 import Color from "terriajs-cesium/Source/Core/Color";
 import TimeIntervalCollection from "terriajs-cesium/Source/Core/TimeIntervalCollection";
+import ConstantPositionProperty from "terriajs-cesium/Source/DataSources/ConstantPositionProperty";
+import ConstantProperty from "terriajs-cesium/Source/DataSources/ConstantProperty";
 import Entity from "terriajs-cesium/Source/DataSources/Entity";
 import PointGraphics from "terriajs-cesium/Source/DataSources/PointGraphics";
+import PropertyBag from "terriajs-cesium/Source/DataSources/PropertyBag";
 import HeightReference from "terriajs-cesium/Source/Scene/HeightReference";
 import filterOutUndefined from "../Core/filterOutUndefined";
 import { JsonObject } from "../Core/Json";
 import TableColumn from "./TableColumn";
 import TableStyle from "./TableStyle";
-import PropertyBag from "terriajs-cesium/Source/DataSources/PropertyBag";
-import ConstantPositionProperty from "terriajs-cesium/Source/DataSources/ConstantPositionProperty";
-import ConstantProperty from "terriajs-cesium/Source/DataSources/ConstantProperty";
 
 type RequiredTableStyle = TableStyle & {
   longitudeColumn: TableColumn;
@@ -23,15 +23,22 @@ export default function createLongitudeLatitudeFeaturePerRow(
   const longitudes = style.longitudeColumn.valuesAsNumbers.values;
   const latitudes = style.latitudeColumn.valuesAsNumbers.values;
   const colorColumn = style.colorColumn;
-  const valueFunction =
+  const colorValueFunction =
     colorColumn !== undefined ? colorColumn.valueFunctionForType : () => null;
+  const pointSizeColumn = style.pointSizeColumn;
+  const pointSizeValueFunction =
+    pointSizeColumn !== undefined
+      ? pointSizeColumn.valueFunctionForType
+      : () => null;
   const colorMap = style.colorMap;
   const pointSizeMap = style.pointSizeMap;
   const outlineColor = Color.fromCssColorString(
     "black" //this.terria.baseMapContrastColor;
   );
   const tableColumns = style.tableModel.tableColumns;
-  const intervals = style.timeIntervals || [];
+  const intervals = style.moreThanOneTimeInterval
+    ? style.timeIntervals ?? []
+    : [];
   const rowIds = style.tableModel.rowIds;
   return filterOutUndefined(
     rowIds.map(rowId => {
@@ -40,15 +47,16 @@ export default function createLongitudeLatitudeFeaturePerRow(
       if (longitude === null || latitude === null) {
         return;
       }
-      const value = valueFunction(rowId);
+      const colorValue = colorValueFunction(rowId);
+      const pointSizeValue = pointSizeValueFunction(rowId);
       const feature = new Entity({
         position: new ConstantPositionProperty(
           Cartesian3.fromDegrees(longitude, latitude, 0.0)
         ),
         point: new PointGraphics({
-          color: new ConstantProperty(colorMap.mapValueToColor(value)),
+          color: new ConstantProperty(colorMap.mapValueToColor(colorValue)),
           pixelSize: new ConstantProperty(
-            pointSizeMap.mapValueToPointSize(value)
+            pointSizeMap.mapValueToPointSize(pointSizeValue)
           ),
           outlineWidth: new ConstantProperty(1),
           outlineColor: new ConstantProperty(outlineColor),
@@ -64,7 +72,7 @@ export default function createLongitudeLatitudeFeaturePerRow(
   );
 }
 
-function getRowValues(
+export function getRowValues(
   index: number,
   tableColumns: Readonly<TableColumn[]>
 ): JsonObject {